requirements.txt
crewai
#crewai_tools
load_dotenv
fastapi
uvicorn
PyPDF2
pandas
python-docx
pdfplumber
python-pptx
SpeechRecognition
moviepy
openpyxl
pydub
beautifulsoup4
markdown
pytube
#wave
#psycopg2
docx2txt
sqlalchemy
sqlalchemy.orm
datetime
typing
pathlib
python-pptx
python-multipart
pymysql
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
tweepy
logging
apscheduler
langchain-openai
playwright
SQLAlchemy
gensim
bertopic
scikit-learn
nltk
numpy
browser-use==0.1.40
#mysqlclient
paramiko
